# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: REST API for Todo (task) CRUD operations backed by SQLite.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
tags:
  - name: tasks
    description: Todo item operations
paths:
  /tasks:
    get:
      tags: [tasks]
      summary: List all tasks
      operationId: listTasks
      responses:
        "200":
          description: Array of tasks ordered by id ascending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [tasks]
      summary: Create a task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          description: Validation or bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    put:
      tags: [tasks]
      summary: Update a task
      description: >
        Provide at least one of `title` or `completed`. Omitting both returns 400.
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          description: Invalid id, validation error, or empty body object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [tasks]
      summary: Delete a task
      operationId: deleteTask
      responses:
        "204":
          description: Task deleted (no response body)
        "400":
          description: Invalid task id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      description: Numeric task identifier (positive integer, no leading zeros)
      schema:
        type: integer
        minimum: 1
  schemas:
    Todo:
      type: object
      required: [id, title, completed, createdAt]
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Buy milk
        completed:
          type: boolean
        createdAt:
          type: string
          description: ISO 8601 timestamp
          example: "2026-04-18T12:00:00.000Z"
    CreateTodoRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        completed:
          type: boolean
          description: Defaults to false if omitted
    UpdateTodoRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
        completed:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message (generic for 500 responses)
        details:
          description: Optional Joi validation details or other diagnostic info
          nullable: true